requirements.txt
# The essential libraries for the project
git+https://github.com/Rapptz/discord.py
asyncio
aiohttp
python-dotenv

# These libraries are for some special commands
requests
better_profanity
mcstatus
psutil
simpcalci
Tabulate
Pillow
Colorama
PyNaCl
Wavelink

# This is for the database
apscheduler

# This library is an addon for discord.py
git+https://github.com/Rapptz/discord-ext-menus


# These are for the dashboard
Quart
Quart_discord
git+https://github.com/ArikSquad/discord-ext-ipc
